Better place for testCreateProjectCheckGoodName()
